renamed auth_application to application
